# Bobeta.Client/Services/Base/NSwag/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Bobeta API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/GamePlay/start": {
      "post": {
        "tags": [
          "GamePlay"
        ],
        "summary": "Starts a match: deals 4 cards each plus stock pile; random first leader.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/GamePlay/play-card": {
      "post": {
        "tags": [
          "GamePlay"
        ],
        "summary": "Plays a card in the current turn. Server validates turn and follow-suit rule, then broadcasts state and move via SignalR.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.PlayCardRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.PlayCardRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.PlayCardRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.GameStateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.GameStateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.GameStateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/GamePlay/void-follow": {
      "post": {
        "tags": [
          "GamePlay"
        ],
        "summary": "Responder holds no cards of the led suit: leader's card is returned, responder draws stock, leader opens again.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.GameStateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.GameStateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.GameStateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/GamePlay/state": {
      "get": {
        "tags": [
          "GamePlay"
        ],
        "summary": "Gets the current game state for the authenticated player (hand, last card, whose turn, game over, winner).",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.GameStateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.GameStateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.GameStateDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Bobeta.Application.DTOs.Game.CardPlayDto": {
        "type": "object",
        "properties": {
          "suit": {
            "$ref": "#/components/schemas/Bobeta.Domain.Enums.CardSuit"
          },
          "rank": {
            "$ref": "#/components/schemas/Bobeta.Domain.Enums.CardRank"
          }
        },
        "additionalProperties": false
      },
...
      "Bobeta.Application.DTOs.Game.GameStateDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "myCards": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "lastPlayedCard": {
            "type": "string",
            "nullable": true
          },
          "currentTurnPlayerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "gameOver": {
            "type": "boolean"
          },
          "winnerPlayerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "waitingForGameStart": {
            "type": "boolean"
          },
          "lobbyPotAmount": {
            "type": "number",
            "format": "double"
          },
          "opponentDisplayName": {
            "type": "string",
            "nullable": true
          },
          "lastTrickWinnerPlayerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "myRoundWins": {
            "type": "integer",
            "format": "int32"
          },
          "opponentRoundWins": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "Bobeta.Application.DTOs.Game.PlayCardRequest": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "card": {
            "$ref": "#/components/schemas/Bobeta.Application.DTOs.Game.CardPlayDto"
          }
        },
        "additionalProperties": false
      },
...
      "Bobeta.Domain.Enums.CardRank": {
        "enum": [
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12,
          13,
          14
        ],
        "type": "integer",
        "format": "int32"
      },
      "Bobeta.Domain.Enums.CardSuit": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Bearer token for Bobeta API.",
        "scheme": "Bearer"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}